Renamed dashboard-files since it will be the title in sidebar, created resources-folder and put in charts.py, png. Changed import for png in top_100.
