+++ src/dashboards/Home.py
@@ -1,0 +1,8 @@
+import streamlit as st
+
+
+st.set_page_config(
+    page_title="Kryptokollen Home",
+)
+
+st.title("Welcome to the Kryptokollen")
