+++ src/dashboards/pages/1_ETH_&_UFD.py
@@ -71,7 +71,6 @@ def layout():
     with col4:
         st.line_chart(df_ufd[price_col], height=300)
 
-
     with st.expander("Detailed ETH Data"):
         st.dataframe(df_eth.tail(10))
